chore: renamed test/util/ to test/fixture/; tap will ignore now
